requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astroid==2.3.0
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.8.1
cachetools==5.3.3
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.3
comm==0.2.2
cryptography==3.3.1
debugpy==1.8.8
decorator==5.1.1
dj-database-url==2.3.0
Django==4.2.4
django-bootstrap-modal-forms==1.5.0
django-bootstrap4==1.1.1
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-heroku==0.3.1
django-rest-knox==5.0.2
django-misaka==0.2.1
django-pdb==0.6.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
executing==2.1.0
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.0
gunicorn==20.0.4
h11==0.14.0
h5py==3.10.0
houdini.py==0.1.0
idna==2.8
ipython==8.29.0
isort==4.3.21
jedi==0.19.2
joblib==1.3.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.14.0
lazy-object-proxy==1.4.3
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.6.1
misaka==2.1.1
ml-dtypes==0.2.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==6.1.0
psycopg2==2.9.7
pure_eval==0.2.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.20
pygeoif==0.7
Pygments==2.18.0
pylint==2.4.4
python-dateutil==2.9.0.post0
python-decouple==3.3
python-dotenv==0.15.0
pytz==2025.1
pyzmq==26.2.0
requests==2.22.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
sgp4==1.4
six==1.17.0
soupsieve==1.9.5
sqlparse==0.4.4
stack-data==0.6.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.3.0
tle2czml==0.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2023.3
urllib3==1.25.7
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.0.1
whitenoise==5.0.1
wincertstore==0.2
wrapt==1.11.2
